+++ example/lib/top_down_game/soldier_player.dart
@@ -68,7 +68,7 @@ class SoldierPlayer extends RotationPlayer with ObjectCollision, Lighting {
           size: Vector2(4, 8),
           speed: 500,
           animationUp: Sprite.load('bullet.png').toAnimation(),
-          damage: 10,
+          damage: 30,
         );
       }
     }